+++ syncany-daemon/src/main/java/org/syncany/operations/daemon/WebServer.java
@@ -34,16 +34,30 @@ import io.undertow.websockets.core.WebSocketChannel;
 import io.undertow.websockets.core.WebSockets;
 
 import java.io.File;
+import java.security.KeyPair;
+import java.security.KeyStore;
+import java.security.cert.Certificate;
+import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.regex.Pattern;
 
+import javax.net.ssl.SSLContext;
+
+import org.bouncycastle.asn1.x500.RDN;
+import org.bouncycastle.asn1.x500.X500Name;
+import org.bouncycastle.asn1.x500.style.BCStyle;
+import org.bouncycastle.asn1.x500.style.IETFUtils;
+import org.bouncycastle.cert.jcajce.JcaX509CertificateHolder;
+import org.syncany.config.UserConfig;
 import org.syncany.config.to.DaemonConfigTO;
 import org.syncany.config.to.UserTO;
+import org.syncany.config.to.WebServerTO;
+import org.syncany.crypto.CipherParams;
+import org.syncany.crypto.CipherUtil;
 import org.syncany.operations.daemon.auth.MapIdentityManager;
 import org.syncany.operations.daemon.handlers.InternalRestHandler;
 import org.syncany.operations.daemon.handlers.InternalWebInterfaceHandler;
@@ -67,7 +81,6 @@ import com.google.common.eventbus.Subscribe;
  */
 public class WebServer {
 	private static final Logger logger = Logger.getLogger(WebServer.class.getSimpleName());
-	public static final Pattern WEBSOCKET_ALLOWED_ORIGIN_HEADER = Pattern.compile("^https?://(localhost|127\\.\\d+\\.\\d+\\.\\d+):\\d+$");
 
 	private Undertow webServer;
 	private LocalEventBus eventBus;
@@ -78,7 +91,7 @@ public class WebServer {
 	
 	private List<WebSocketChannel> clientChannels;
 
-	public WebServer(DaemonConfigTO daemonConfig) {
+	public WebServer(DaemonConfigTO daemonConfig) throws Exception {
 		this.clientChannels = new ArrayList<WebSocketChannel>();
 		
 		initCaches();
@@ -116,38 +129,109 @@ public class WebServer {
 		eventBus.register(this);
 	}
 
-	private void initServer(DaemonConfigTO daemonConfigTO) {
-		String host = daemonConfigTO.getWebServer().getHost();
-		int port = daemonConfigTO.getWebServer().getPort();
+	private void initServer(DaemonConfigTO daemonConfigTO) throws Exception {
+		WebServerTO webServerConfig = daemonConfigTO.getWebServer();
+		
+		// Bind address and port
+		String bindAddress = webServerConfig.getBindAddress();
+		int bindPort = webServerConfig.getBindPort();
+
+		// Users (incl. CLI user!)
+		List<UserTO> users = readWebServerUsers(daemonConfigTO);
+		IdentityManager identityManager = new MapIdentityManager(users);
+		
+		// (Re-)generate keypair/certificate (if requested)
+		boolean certificateAutoGenerate = webServerConfig.isCertificateAutoGenerate();
+		String certificateCommonName = webServerConfig.getCertificateCommonName();
+		
+		if (certificateAutoGenerate && certificateCommonNameChanged(certificateCommonName)) {
+			generateNewKeyPairAndCertificate(certificateCommonName);	
+		}		
+		
+		// Set up the handlers for WebSocket, REST and the web interface		
+		HttpHandler pathHttpHandler = path()
+			.addPrefixPath("/api/ws", websocket(new InternalWebSocketHandler(this, certificateCommonName)))
+			.addPrefixPath("/api/rs", new InternalRestHandler(this))
+			.addPrefixPath("/", new InternalWebInterfaceHandler());
+		
+		// Add some security spices
+		HttpHandler securityPathHttpHandler = addSecurity(pathHttpHandler, identityManager);
+		SSLContext sslContext = UserConfig.createUserSSLContext();
+		
+		// And go for it!
+		webServer = Undertow
+			.builder()
+			.addHttpsListener(bindPort, bindAddress, sslContext)
+			.setHandler(securityPathHttpHandler)
+			.build();
+		
+		logger.log(Level.INFO, "Initialized web server.");
+	}
+
+	private List<UserTO> readWebServerUsers(DaemonConfigTO daemonConfigTO) {
 		List<UserTO> users = daemonConfigTO.getUsers();
 		
 		if (users == null) {
 			users = new ArrayList<UserTO>();
 		}
 		
+		// Add CLI credentials
 		if (daemonConfigTO.getPortTO() != null) {
-			// Add CLI credentials
 			users.add(daemonConfigTO.getPortTO().getUser());
 		}
 		
-		IdentityManager identityManager = new MapIdentityManager(users);
-		
-		HttpHandler pathHttpHandler = path()
-			.addPrefixPath("/api/ws", websocket(new InternalWebSocketHandler(this)))
-			.addPrefixPath("/api/rs", new InternalRestHandler(this))
-			.addPrefixPath("/", new InternalWebInterfaceHandler());
-			
-		HttpHandler securityPathHttpHandler = addSecurity(pathHttpHandler, identityManager);
-		
-		webServer = Undertow
-			.builder()
-			.addHttpListener(port, host)
-			.setHandler(securityPathHttpHandler)
-			.build();
-		
-		logger.log(Level.INFO, "Initialized websocket server.");
+		return users;
 	}
 	
+	private boolean certificateCommonNameChanged(String certificateCommonName) {
+		try { 
+			KeyStore userKeyStore = UserConfig.getUserKeyStore();
+			X509Certificate currentCertificate = (X509Certificate) userKeyStore.getCertificate(CipherParams.CERTIFICATE_IDENTIFIER);
+			
+			if (currentCertificate != null) {
+				X500Name currentCertificateSubject = new JcaX509CertificateHolder(currentCertificate).getSubject();
+				RDN currentCertificateSubjectCN = currentCertificateSubject.getRDNs(BCStyle.CN)[0];
+
+				String currentCertificateSubjectCnStr = IETFUtils.valueToString(currentCertificateSubjectCN.getFirst().getValue());
+				
+				if (!certificateCommonName.equals(currentCertificateSubjectCnStr)) {
+					logger.log(Level.INFO, "- Certificate regeneration necessary: Cert common name in daemon config changed from " + currentCertificateSubjectCnStr + " to " + certificateCommonName + ".");
+					return true;
+				}				
+			}
+			else {
+				logger.log(Level.INFO, "- Certificate regeneration necessary, because no certificate found in key store.");
+				return true;
+			}			
+			
+			return false;
+		}
+		catch (Exception e) {
+			throw new RuntimeException("Cannot (re-)generate server certificate for hostname: " + certificateCommonName, e);		
+		}
+	}
+
+	public static void generateNewKeyPairAndCertificate(String certificateCommonName) {
+		try {
+			logger.log(Level.INFO, "(Re-)generating keypair and certificate for hostname " + certificateCommonName + " ...");
+			
+			// Generate key pair and certificate
+			KeyPair keyPair = CipherUtil.generateRsaKeyPair();
+			X509Certificate certificate = CipherUtil.generateSelfSignedCertificate(certificateCommonName, keyPair);
+			
+			// Add key and certificate to key store
+			UserConfig.getUserKeyStore().setKeyEntry(CipherParams.CERTIFICATE_IDENTIFIER, keyPair.getPrivate(), new char[0], new Certificate[] { certificate });
+			UserConfig.storeUserKeyStore();
+			
+			// Add certificate to trust store (for CLI->API connection)
+			UserConfig.getUserTrustStore().setCertificateEntry(CipherParams.CERTIFICATE_IDENTIFIER, certificate);
+			UserConfig.storeTrustStore();
+		}
+		catch (Exception e) {
+			throw new RuntimeException("Unable to read key store or generate self-signed certificate.", e);
+		}
+	}		
+
 	private static HttpHandler addSecurity(final HttpHandler toWrap, IdentityManager identityManager) {		
 		List<AuthenticationMechanism> mechanisms = 
 				Collections.<AuthenticationMechanism> singletonList(new BasicAuthenticationMechanism("Syncany"));
@@ -160,7 +244,7 @@ public class WebServer {
 		handler = new SecurityInitialHandler(AuthenticationMode.PRO_ACTIVE, identityManager, handler);
 		
 		return handler;
-	}
+	}	
 
 	private void sendBroadcast(String message) {
 		synchronized (clientChannels) {